+++ lib/resque/kubernetes.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 require "resque/kubernetes/configurable"
+require "resque/kubernetes/context/kubectl"
+require "resque/kubernetes/context/well_known"
 require "resque/kubernetes/context_factory"
 require "resque/kubernetes/deep_hash"
 require "resque/kubernetes/dns_safe_random"
@@ -19,6 +21,9 @@ module Resque
 
     # Limit the number of workers that should be spun up, default 10
     define_setting :max_workers, 10
+
+    # A `kubeclient` for connection context, default attempts to read from cluster or `~/.kube/config`
+    define_setting :kubeclient, nil
   end
 end