+++ spec/resque/kubernetes/job_spec.rb
@@ -58,7 +58,8 @@ describe Resque::Kubernetes::Job do
   let(:jobs_client) { spy("jobs client") }
 
   before do
-    allow_any_instance_of(Resque::Kubernetes::JobsManager).to receive(:jobs_client).and_return(jobs_client)
+    allow(Kubeclient::GoogleApplicationDefaultCredentials).to receive(:token).and_return("token")
+    allow(Kubeclient::Client).to receive(:new).and_return(jobs_client)
   end
 
   shared_examples "before enqueue callback" do
@@ -68,13 +69,30 @@ describe Resque::Kubernetes::Job do
       let(:done_pod)    { K8sStub.new(status: {phase: "Succeeded"}) }
       let(:working_pod) { K8sStub.new(status: {phase: "Running"}) }
 
+      context "when Resque::Kubernetes.kubeclient is defined" do
+        let(:client) { spy("custom client") }
+
+        before do
+          allow(Kubeclient::Client).to receive(:new).and_call_original
+          allow(Resque::Kubernetes).to receive(:kubeclient).and_return(client)
+        end
+
+        it "uses the provided client" do
+          expect(client).to receive(:get_jobs).at_least(:once).and_return([])
+          expect(client).to receive(:create_job)
+          subject.before_enqueue_kubernetes_job
+        end
+      end
+
       context "when Rails.env is not defined" do
         before do
           expect(defined? Rails).not_to be true
         end
 
         it "calls kubernetes APIs" do
-          expect_any_instance_of(Resque::Kubernetes::JobsManager).to receive(:jobs_client).and_return(jobs_client)
+          expect_any_instance_of(Resque::Kubernetes::JobsManager).to receive(:jobs_client).at_least(:once) do
+            jobs_client
+          end
           subject.before_enqueue_kubernetes_job
         end
       end
@@ -93,7 +111,9 @@ describe Resque::Kubernetes::Job do
           end
 
           it "calls kubernetes APIs" do
-            expect_any_instance_of(Resque::Kubernetes::JobsManager).to receive(:jobs_client).and_return(jobs_client)
+            expect_any_instance_of(Resque::Kubernetes::JobsManager).to receive(:jobs_client).at_least(:once) do
+              jobs_client
+            end
             subject.before_enqueue_kubernetes_job
           end
         end
@@ -321,15 +341,42 @@ describe Resque::Kubernetes::Job do
             end
           end
 
-          context "when the namespace is not included" do
-            it "sets it to 'default'" do
-              manifest = hash_including(
-                  "metadata" => hash_including(
-                      "namespace" => "default"
-                  )
-              )
-              expect(Kubeclient::Resource).to receive(:new).with(manifest).and_return(job)
-              subject.before_enqueue_kubernetes_job
+          context "when the namespace is not included in the manifest" do
+            context "and no value is provided by the authentication context" do
+              it "sets it to 'default'" do
+                manifest = hash_including(
+                    "metadata" => hash_including(
+                        "namespace" => "default"
+                    )
+                )
+                expect(Kubeclient::Resource).to receive(:new).with(manifest).and_return(job)
+                subject.before_enqueue_kubernetes_job
+              end
+            end
+
+            context "and the authentication context provides a namespace" do
+              let(:context) do
+                OpenStruct.new(
+                    endpoint:  "https://127.0.0.0",
+                    version:   "v1",
+                    namespace: "space",
+                    options:   {}
+                )
+              end
+
+              before do
+                allow(Resque::Kubernetes::ContextFactory).to receive(:context).and_return(context)
+              end
+
+              it "uses the context-provided namespace" do
+                manifest = hash_including(
+                    "metadata" => hash_including(
+                        "namespace" => "space"
+                    )
+                )
+                expect(Kubeclient::Resource).to receive(:new).with(manifest).and_return(job)
+                subject.before_enqueue_kubernetes_job
+              end
             end
           end